add LICENSE file (#15)
I'm looking at packaging this crate for Fedora Linux as a dependency of
[ruff](https://github.com/astral-sh/ruff), and noticed this project
doesn't include the license text for the license that's specified in
crate metadata.

The MIT license is one of the licenses that requires a copy of the
license text to be included in (re)distributed copies of projects'
source code, so missing this file presents a potential problem.

The file added with this PR is a copy of the MIT license text from the
rust-lang/rust repository (it is a 100% match for what SPDX defines as
the "MIT" license).

If this PR is acceptable, it would be great to get a new version
released to crates.io with the license file included. Thank you!